Move updates per second constant to Simulator class from Histogram class and change type of time elapsed member variable in Histogram class to double instead of size_t

+++ include/core/histogram.h
@@ -22,11 +22,10 @@ class Histogram {
   const double kHistogramGraphDimension = 400;  // length and width of histogram graph
   //const size_t kLabelSpacingFromHistogram = 8;
   const size_t kSpacingFromContainer = 125;
-  //const size_t kUpdatesPerSecond = 50;
 
   double upper_bound_for_y_;  // i.e. the highest label value for y axis
   vec2 container_top_right_corner_;
-  size_t time_elapsed_since_outbreak_;
+  double time_elapsed_since_outbreak_;
 
   // the key is a Status representing the health status
   // the value is a vector representing the people with the specified health status
